requirements.txt
Django==4.2.14
django-filter==24.3
requests==2.32.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.0.0
sympy==1.13
djangorestframework-api-key==2.1.0
celery==5.3.4
django-celery-beat==2.5.0
django-redis==5.3.0
django-extensions==3.2.1
importlib-metadata==6.8.0
django-debug-toolbar==4.1.0
